[Bronze V] Title: 이상한 기호, Time: 204 ms, Memory: 17628 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,8 +3,8 @@ import java.math.BigInteger;
 public class Main {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
-        int a = s.nextInt();
-        int b = s.nextInt();
+        long a = s.nextInt();
+        long b = s.nextInt();
         System.out.print((a+b) * (a-b));
     }
 }
